+++ CHANGELOG.md
@@ -11,3 +11,7 @@
 
 ## [0.1.1+2] - 2025-03-15
 - Fix for linting
+- 
+## [0.1.2] - 2025-03-20
+- Add feature for issue #2: add momentum scrolling
+- Add feature for issue #3: allow zooming in/out with mouse wheel